global追加
