+++ AutoMouse/SetParam.cpp
@@ -10,9 +10,10 @@
 #include "GnrlDefine.h"
 
 #include "GnrlFilepath.h"
-#include "AM_define.h"
-#include "AM_extern.h"
-#include "ProcBase.h"
+#include "define.h"
+#include "extern.h"
+#include "global.h"
+#include "WorkBase.h"
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
 // XML ファイル読み込み
@@ -22,7 +23,9 @@ int32_t initParam()
 	int32_t ack;
 	TCHAR InitFilePath[MAX_PATH];
 
-	ack = GnrlFilepath::getModuleAttachmentFilePath(InitFilePath, MAX_PATH ,_T("am_ini.xml"));
+	// **** 設定ファイルオープン ****
+
+	ack = GnrlFilepath::getModuleAttachmentFilePath(InitFilePath, MAX_PATH ,_T("apd_ini.xml"));
 	if(ack <= 0)	InitFilePath[0] = '\x0';
 
 	rapidxml::document_t doc;
@@ -31,13 +34,25 @@ int32_t initParam()
 	ack = rapidxml::load_document(doc ,InitFilePath ,docbuf);
 	if(ack < 0)	return ERC_ng;
 
+	// **** 各種パラメータ読み込み ****
+
 	rapidxml::node_t *root = doc.first_node();
+	rapidxml::node_t *node;
 	rapidxml::first_node(root ,_T("title") ,gTitle);
+
+	node = rapidxml::first_node(root, _T("bure"));
+	if (node == nullptr){
+		gBurePoint = CPoint(4, 4);
+	}else{
+		int32_t val;
+		rapidxml::first_attribute(node, _T("x"), val);	gBurePoint.x = val;
+		rapidxml::first_attribute(node, _T("y"), val);	gBurePoint.y = val;
+	}
 
 	rapidxml::node_t *work	= rapidxml::first_node(root ,_T("works"));
 	if(work == nullptr)	return ERC_ng;
 
-	ack = ProcBase::loadProcList(gWorks ,work);
+	ack = WorkBase::loadProcList(gWorks, work, gWorkNames);
 
 	return ERC_ok;
 }
