+++ AutoMouse/sub.h
@@ -1,0 +1,15 @@
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+//  sub.h
+// 
+// 
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+
+#include <stdafx.h>
+
+
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+// ƒ‰ƒ“ƒ_ƒ€
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+int random(int MaxVal);
+int randScatter(int CenterVal, int Width);
+
